Fix Intel eDP refresh rate issue
This resolves T6025 by applying the upstream patch to ensure that 60hz
is still respected.

Note as 4.15.x is now EOL we have not updated to the latest point release
as it doesn't cover any security issues, instead we will test a branch
jump to 4.16.x next week.

Maniphest Tasks: T6025

Signed-off-by: Ikey Doherty <ikey@solus-project.com>